Set up the Router
We are using React Router whenever we are in a certain path="", we would like to render a certain component. React Router uses the Link component, whenever we click one of the link with the path, it triggers the Router to render a certain component. We also passes the state to the component we are rendering