Build ContactPage Component
We edited the ContactPage component to a class component.
We added the components state name, email, message, disabled and emailSent.
We created 2 functions that handles changes to the Textboxes and the Submit Button.
We used boostrap Form component to create our Contact page form, and Submit button to handle sending the Form to the server